Project Importing
This is an application example that is going to be used as tutorial for a JavaScript Refactoring course. On this commit you will be able to find JS code that is frequently used on production Web Apps. It contains a lot of bad practices and it pretty easy to find code smells on it.